Extracted UI KIT as separate component.
