+++ src/ignitus-AboutPage/ignitus-About/Components/About.tsx
@@ -4,7 +4,7 @@ import CoreTeam from '../../ignitus-CoreTeam';
 import {withErrorBoundary} from '../../../ignitus-Shared/Components/errorBoundary';
 import * as t from '../constants';
 import * as S from '../Styles';
-import * as T from '../../../ignitus-Shared/Components/ignitus-DesignSystem/shared';
+import * as T from '../../../ignitus-Shared/ignitus-DesignSystem/shared';
 
 const About = () => (
   <Fragment>
