+++ src/ignitus-AboutPage/ignitus-CoreTeam/Components/team.tsx
@@ -5,7 +5,7 @@ import React from 'react';
 import {Data} from '../constants';
 import {withErrorBoundary} from '../../../ignitus-Shared/Components/errorBoundary';
 import * as S from '../Styles';
-import * as T from '../../../ignitus-Shared/Components/ignitus-DesignSystem/shared';
+import * as T from '../../../ignitus-Shared/ignitus-DesignSystem/shared';
 import {coreTeam, TeamItemProps} from '../types';
 
 const PureCoreTeam = ({team}: any) => (
