+++ src/ignitus-HomePage/ignitus-Contributors/Components/Contributors.tsx
@@ -3,7 +3,7 @@ import React from 'react';
 import { Data } from '../constants';
 import { withErrorBoundary } from '../../../ignitus-Shared/Components/errorBoundary';
 import * as S from '../Styles';
-import * as T from '../../../ignitus-Shared/Components/ignitus-DesignSystem/shared';
+import * as T from '../../../ignitus-Shared/ignitus-DesignSystem/shared';
 
 interface Props {
   data: any;
