+++ src/ignitus-Shared/ignitus-DesignSystem/ignitus-Templates/ignitus-AuthenticationComponents/Login/Login.jsx
@@ -6,8 +6,8 @@ import { Redirect, Link } from 'react-router-dom';
 import isEmpty from 'lodash/isEmpty';
 import PropTypes from 'prop-types';
 import * as t from './Constants';
-import loader from '../../../../../../ignitus-Assets/Images/loader2.gif';
-import { withErrorBoundary } from '../../../../errorBoundary';
+import loader from '../../../../../ignitus-Assets/Images/loader2.gif';
+import { withErrorBoundary } from '../../../../Components/errorBoundary';
 
 const Login = ({
   loginType, tagline, handleSubmit, logInData, state, setState,
