+++ src/ignitus-Shared/ignitus-DesignSystem/ignitus-Templates/ignitus-DashboardSideNavigation/style.scss
@@ -1,5 +1,5 @@
-@import '../../../../styles/ignitus-components-common.scss';
-@import '../../../../styles/ignitus-tokens.scss';
+@import '../../../styles/ignitus-components-common.scss';
+@import '../../../styles/ignitus-tokens.scss';
 
 .navigator {
   position: absolute;
