+++ src/ignitus-UserInterfaceKit/styles.ts
@@ -1,8 +1,8 @@
 
 import styled from '@emotion/styled';
-import * as A from '../ignitus-DesignSystem/ignitus-Atoms/atoms';
-import * as C from '../ignitus-DesignSystem/ignitus-Atoms/colors';
-import * as F from '../ignitus-DesignSystem/ignitus-Atoms/fonts';
+import * as A from '../ignitus-Shared/ignitus-DesignSystem/ignitus-Atoms/atoms';
+import * as C from '../ignitus-Shared/ignitus-DesignSystem/ignitus-Atoms/colors';
+import * as F from '../ignitus-Shared/ignitus-DesignSystem/ignitus-Atoms/fonts';
 
 export const UIKitContainer = styled.div`
   padding: 5em;
